refactor(home): リファクタリング
